Security and logic fixes.
Someone could ostensibly enter an escape in raw_input and foobar the .todo
directory. Also I had an if statement like
if not a or not b:
    ...
which is more simply put
if not (a and b):
    ...
Also importing the readline module, which I have to fix for OSX users, now
that I think about it.