Fix a couple minor errors. Add lsc | listcon.
There was a minor problem when appending to lines. A gratuitous space was
added in between the last character and the \n. I also "fixed" the behavior
when running './todo.py (pri|pre|app)' so that those functions don't need to
end with sys.exit(0). Also change the format for dates. I had forgotten that
contexts used @'s in the traditional (read: original) version.