Check that todo.py is not already installed.
If it is already installed check that there is both a difference in the file
itself and that our version is newer. It could be the case that the user has
modified the file in their $INSTALL_DIR and they don't want the old version to
overwrite it (they should be editing their github fork and working there, but
just in case they aren't).

I'd rather be safe than have an annoyed user who idiotically ran the script
without realizing it would cancel out his work.