Simplification (readability) add 'a' as a command.
I was browsing over the code to re-familiarize myself with it (since
winter break is coming up*) and noticed in prompt() that I was using a
list comprehension when simply calling list() would be easier to
understand (especially for someone just learning python). I will have to
admit that I find list comprehensions absolutely elegant and love to
abuse the use of them.

Also, I noticed jvstein added the 'a' alias for the 'add' command. He
hasn't issued a pull request so I thought I'd add it anyway (I'm a bit
impatient).

*by which I mean, "Now I'll have a bit of time to work on this some
more"