Change the way printing occurs.
This makes grep'ing the output far easier. For issue #8 I'm going to eventually
to have a single script to run all of them, hence the individual main()'s for
each file.

And right now I'm only doing 11 for the sake of speed. I also want to eventually
test the way _list_ works. Which is why I'm considering the double digits. This
will ensure nothing happens to the padding patch that @jvstein contributed a
short while ago. It should also test for ordering, etc. That'll probably be the
most... intensive of the functions.