Perfect 'pri' command. Add more tests.
While writing the tests for the prioritize command, I realized that it would
actually accept anything and everything as a priority. Someone could have tried
~$ todo.py pri 3 ABCDEFGHIJKLMNOPQRSTUVWXYZ
And it probably would have worked. Now that command first checks to make sure
it's a single letter and then if it's in the list of priorities. Only then will
it add the given priority to to the line.