Added: modelines, assertIsInstance[1], format[2]
[1] assertIsInstance is not in Python 2.6, so I hacked together a fix for
that into base.py (never do I think I have ever appreciated OOP more)

[2] I feel confident that what I have written is correct. I'll have to write
a fix for assertIsNotNone for Python 2.6, but until then, the code written
there works and is fine. I expect to be using the IsNotNone assertion
frequently in this file and it might be useful to have it in the base.

Modelines are just so I have no indentation errors in (g)vim.

Issue #8 for reference.