:39s/Name/Attribute/ & initialize formatted better
I was catching the wrong error when trying to alias sys.intern as intern for
python3 users => :39s/Name/Attribute

In the format_lines() function, I had a really convoluted way of initializing
the formatted dictionary (that is if we weren't only doing colored lines). I
realized there was an easier one-liner that will do it for me. In short, for
some reason I didn't realize that a comprehension like:
[[] for i in PRIORITIES] would probably be faster than running map on a function
that simply initializes each key in the dictionary with an empty list. Also,
this is probably far more obvious to a new-comer since composing dict() with
zip() is a pretty popular trick in books that teach python.

Also, this works fine for python3.