Total PEP8 compliance and small optimizations.
@265: dir is a built-in function and although python's namespaces are more
intelligent, I don't want to step on toes.

@777: switch out three lines for one. default is going to be reset regardless,
so using it to store a value from the dictionary makes sense. This allows for a
short ternary line.
Also, for PEP8 compliance there was a rather long line (82 chars) that involved
TERM_COLORS.keys() and I realized that that was being called every time it was
checking for the value in the list. Since it makes more sense to get those
values once (since the dictionary isn't going to change) and sets are apparently
faster when checking for inclusion, I assign the set with the value of
TERM_COLORS.keys() to colors and use that.

@799: I would hope that using the same variable will produce some slight optimization.

@865: The construct to me was horribly ugly. This makes sense since I'm only
running compile once in the worst case scenario, so since I know the re module
is the largest culprit in terms of performance, this should provide some time
savings in a frequently used function.

@896: I noticed that the matches weren't case insensitive like they are in
todo.sh. So this is a minor behavior modification.

@1059: I don't know why, but I had a space in between "__main__" and :. pep8.py
caught it thankfully. I can't for the life of me figure out why that was like
that.

@1110: Again, the look-up speed of sets is supposed to be faster and this allows
me to trim another line that was too long.