Fix an important bug.
Back when I was using 'dir' as the name of the new directory name, I changed it
so that it would be dir_name instead (to avoid conflicting with the builtin
dir()) but I neglected to set CONFIG["TODO_DIR"] to dir_name instead of dir. So
using a different directory would produce an error with the
os.environ["TODO_DIR'] = CONFIG["TODO_DIR"] line where it was trying to assign a
function object to the environment.