Fix get_config and prioritize_todo.
If someone specifies both a config file and a todo directory, it's
likely they are not in the same directory. The old way of updating the
default values would overwrite the custom config file. Changing the
order should fix this.

In prioritize_todo, if someone only provided './todo.py pri 1' without a
priority, the script will crash. Shortening args to a allows us to stay
within pep8 standards of 80 characters per line or less. And changing
the priority to an upper case will prevent errors when someone uses a
lowercase priority. We should be smarter than that.

Thanks to Matt Stave for pointing these out.